feat: add footer, success component, favicon, and implement the AddParticipant, create platform, and PublishContent methods in the Team page
