+++ app/mediaExperts/[userId]/courses/[courseId]/joinCourse/page.tsx
@@ -1,8 +1,13 @@
-const JoinCoursePage = () => {
+import Success from "../../../../../../components/Success";
+
+const JoinCoursePage = ({ params }: { params: { userId: string } }) => {
   return (
-    <div>
-      <h1>Join Course</h1>
-    </div>
+    <Success
+      title="Course Joined Successfully"
+      description="You have successfully joined the course."
+      buttonText="Go to Dashboard"
+      buttonLink={`/mediaExperts/${params.userId}/admin`}
+    />
   );
   }
