+++ components/Footer.tsx
@@ -1,0 +1,9 @@
+const Footer: React.FC = () => {
+  return (
+    <footer className="p-4 bg-dark-500 text-light-200 text-center">
+      <p>&copy; 2024 Peer Voice. All rights reserved.</p>
+    </footer>
+  );
+};
+
+export default Footer;
